# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sentinel Enforcement Platform API
  version: 0.1.0
  description: Phase 0-5 vertical slice for health, report intake, authorization, case lifecycle, workflow, and storage-backed evidence intake.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/recommendations/{recommendationId}/submit:
    post:
      operationId: submitRecommendation
      tags:
        - recommendations
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: recommendationId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recommendation submitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '401':
          description: Request is unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Authenticated actor is not allowed to submit the recommendation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recommendation not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Recommendation cannot be submitted in its current state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/recommendations/{recommendationId}/reviews:
    post:
      operationId: reviewRecommendation
      tags:
        - recommendations
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: recommendationId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRecommendationRequest'
      responses:
        '200':
          description: Recommendation reviewed and approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '400':
          description: Request body failed validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Request is unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Authenticated actor is not allowed to review the recommendation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recommendation not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Recommendation review violates maker-checker or current state rules.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReviewRecommendationRequest:
      type: object
      required:
        - reviewSummary
      properties:
        reviewSummary:
          type: string
          minLength: 1
          maxLength: 2000
    RecommendationStatusValue:
      type: string
      enum:
        - DRAFT
        - SUBMITTED
        - APPROVED
    RecommendationResponse:
      type: object
      required:
        - id
        - caseId
        - title
        - summary
        - proposedDecision
        - status
        - createdAt
        - createdBy
        - updatedAt
        - updatedBy
        - version
      properties:
        id:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        title:
          type: string
        summary:
          type: string
        proposedDecision:
          type: string
        proposedSanction:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/RecommendationStatusValue'
        submittedAt:
          type: string
          format: date-time
          nullable: true
        submittedBy:
          type: string
          nullable: true
        approvedReviewId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
        version:
          type: integer
          format: int64
...
    ErrorResponse:
      type: object
      required:
        - type
        - title
        - status
        - code
        - detail
        - instance
        - correlationId
        - violations
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        instance:
          type: string
        correlationId:
          type: string
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
    Violation:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string